Revert to using GUILE_LOAD_PATH for guix, but without including in sudoers
The ~/.local/bin script method unfortunately doesn't work due to infinitely
calling itself if I try to make it transparent (i.e. named the same as guix).
Instead, we use GUILE_LOAD_PATH again, but without the potentially dangerous
modification to sudoers.  This has the caveat of requiring that guix be
bootstrapped in someway when reconfiguring the system as root, but it should
be (at least partially) alleviated by upcoming additions of aliases to make it
easier to reconfigure the system.