my-guix: Update mesa-git
This commit updates Mesa to the latest buildable commit as of writing.  Later
commits appear to have a libclc issue that I haven't figured out how to resolve
yet.  There is another unrelated issue that obscures pinpointing which commit
introduced the former issue, although this one has already been fixed.

The libclc input in this package currently does nothing; one of the later
commits adds this as a required dependency which will then be used.

For the purpose of avoiding name collision as an unstable package, this has also
been renamed to mesa-git.