my-guix: mesa-git: Update to commit 944ef17
This commit modifies mesa-git so that it works with recent changes to mesa,
bringing the package back to HEAD.  The intel-clc flag is now enabled and
enforced; this introduces libclc, llvm-15, and clang-15 as inputs.

Due to the i686-linux build being unused currently, we'll switch back to using
Guix's rust.

Patches have also been updated to work with the current version of mesa.