my-guix: Add procedure for patching crate wrap files
This procedure enables wrap files to be patched such that crates packaged in
Guix can be used rather than duplication of effort from downloading sources
separately.  Additionally, this patch script generator offers another
improvement in that versions no longer need to be pinned - whatever's in Guix
can be used (as long as the version is compatible, of course).