my-guix: swapfile-mod: Utilize compose and parameters
Renames build-swapfile-mod -> swapfile-mod.

A swapfile parameter has been added and exported for use with swapfile-mod.

This is a breaking change.  Systems will be updated individually to use
swapfile-mod with parameters as needed.

swapfile-mod has also been updated to no longer be a procedure, using the newer
method of composing mods and utilizing parameters for dynamic configuration
variables.