my-guix: Add conversion procedures for modded-system
This adds procedures that will enable converting a `modded-system' record to
a `home-environment' or an `operating-system'.  Some TODO comments are added for
moving to fields that will eventually be added to `mod', which should allow for
combining `home-environment' and `operating-system' mods/configurations.

Internal changes with mod exclusion have been made.

The `exclude-mods' parameter has been renamed to `excluded-mods' and is now
exported, making it the main way of excluding mods (once `apply-mods' is phased
out).