my-guix: modded-system: Use initial-{os,he} as field names instead of base-{os,he}
Since extensions could be made for stuff like user account groups, it would be
more optimal to prefer as much configuration be made in the initial records -
that mods then extend - rather than doing so after the mods have been applied.
The `base-*' naming appears to encourage the latter, so renaming to `initial-*'
seems reasonable.