my-guix: mod: Add os and he extension fields
This deprecates the apply field in favor of os-extension and he-extension, which
will modify an operating-system and home-environment record, respectively.  This
enables eventually combining the home and system mods (and config files as
well).

The apply field is now used only when it is not the identity procedure; i.e.,
whenever the apply field is not the default.