my-guix: modded-system: Add final-os-extension and final-he-extension
These extensions will make it much easier to work with combined home-system
configurations (when it is eventually supported).  This allows for handling
cases where a user might want to make additional or even temporary
configurations to the "completed" configuration or where ordering is important;
for example, modifying a service added by a mod.