my-guix: Handle more edge cases with wrapped-guix-script
It seems more robust to check for existence of these paths at runtime instead of
during script creation, so the contents of guix-path have been merged (and
gexp'd) into wrapped-guix-script.

GUIX_PACKAGE_PATH is also only set when it is unset, which resolves cases where
one might want to override the value with something else (e.g. "", perhaps for
a pure environment).