files/emacs: project-dispatch: Lazily prompt for project root
It can be unexpected to enter the transient not being aware of whether there is
a prompt for the root directory or not, since one might quickly enter chords
only for the prompt to have appeared.  Instead, prompt for the root directory
only when it is needed so the prompt happens during the suffix call or when the
user explicitly switches the project.