files/emacs: project-dispatch: Improve validation/handling of root directory
This change makes it so that the user-submitted directories in the same project
always derive the same value that "--root-directory" is set to; in other words,
"--root-directory" is always set to the actual project's root directory, even
when a subdirectory is inputted.

Root directories accessed via the transient will always be remembered now.

Since only real project root directories can be set, a message is printed when
it cannot be found.

This logic is also delegated to a function, since :init-value and :reader uses
the same logic under the hood.