emacs: project-dispatch: Add a fallback for project-dispatch--with-environment.
It seems like a good idea to try making 'project-dispatch--with-environment'
usable for all suffixes, regardless of which prefix they are called from.
This means, however, that values are not always guaranteed to exist, so this
fixes an edge case where 'default-directory' gets undesirably set to nil.

* files/emacs/impure/lisp/project-dispatch.el
(project-dispatch--with-environment): Add a fallback to 'default-directory' for
'from-directory'.