emacs: transients: Remove related-projects command
`project-name` requires access to the project files (specifically dir-locals),
so if a project's root belongs to another machine which needs SSH access, a
connection attempt will be made.  This will likely be undesirable.

Project directory basename is thus the only other option - which is not
flexible enough - so we sunset this command for now and come back to it later
in case there is another better solution.